glib2: suppress a warning

+++ glib2/ext/glib2/rbglib_regex.c
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2015-2016  Ruby-GNOME2 Project Team
+ *  Copyright (C) 2015-2019  Ruby-GNOME2 Project Team
  *
  *  This library is free software; you can redistribute it and/or
  *  modify it under the terms of the GNU Lesser General Public
@@ -428,7 +428,8 @@ rg_replace(gint argc, VALUE *argv, VALUE self)
 static VALUE
 rg_s_escape_string(G_GNUC_UNUSED VALUE self, VALUE string)
 {
-    return CSTR2RVAL(g_regex_escape_string(RVAL2CSTR(string), RSTRING_LEN(string)));
+    return CSTR2RVAL(g_regex_escape_string(RVAL2CSTR(string),
+                                           (gint)RSTRING_LEN(string)));
 }
 
 static VALUE
